+++ serie2/lukecreated-serie2/src/components/footer.jsx
@@ -1,4 +1,3 @@
-// import { styled } from "../styles/stitches.config";
 function Footer() {
   return (
     <>