+++ changelog.md
@@ -75,3 +75,4 @@ Created on May 22, 2023
 - Initial Development: Contact dialog is added, and minor improvements on July 31, 2024 -> __0.1.3__
 - Initial Development: Add support for multi paths, and minor improvements on July 31, 2024 -> __0.1.4__
 - Initial Development: Draft for book contents on August 1, 2024 -> __0.1.5__
+- Initial Development: Improved navigation to other page on navbar on August 1, 2024 -> __0.1.6__